+++ pyvela/tests/test_run_analysis.py
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 
 from pyvela.spnta import SPNTA
-from pyvela import pyvela_compare_script, pyvela_script
+from pyvela import pyvela_compare_script, pyvela_script, pyvela_plot_script
 from pint.models import get_model_and_toas
 
 prior_str = """
@@ -16,8 +16,9 @@ prior_str = """
             "args": [0.5, 1.5]
         },
         "EQUAD": {
-            "distribution": "LogUniform",
-            "args": [1e-3, 1.0]
+            "distribution": "Normal",
+            "args": [0.0, 1.0],
+            "lower": 0.0
         }
     }
 """
@@ -58,17 +59,17 @@ def test_analysis_NGC6440E_emcee():
     )
 
 
-def test_script_NGC6440():
+@pytest.mark.parametrize("dataset", ["NGC6440E", "sim_sw.wb"])
+def test_script(dataset):
     datadir = os.path.dirname(os.path.realpath(__file__)) + "/datafiles"
-    dataset = "NGC6440E"
     parfile, timfile = f"{datadir}/{dataset}.par", f"{datadir}/{dataset}.tim"
-    outdir = "_NGC6440E_out"
+    outdir = f"_{dataset}_out"
 
     prior_file = "__prior.json"
     with open(prior_file, "w") as pf:
         print(prior_str, file=pf)
 
-    args = f"{parfile} {timfile} -P {prior_file} -o {outdir}".split()
+    args = f"{parfile} {timfile} -P {prior_file} -T {parfile} -o {outdir}".split()
 
     pyvela_script.main(args)
 
@@ -80,6 +81,8 @@ def test_script_NGC6440():
     assert os.path.isfile(f"{outdir}/param_scale_factors.txt")
     assert os.path.isfile(f"{outdir}/samples_raw.npy")
     assert os.path.isfile(f"{outdir}/samples.npy")
+
+    pyvela_plot_script.main([f"{outdir}/"])
 
 
 @pytest.mark.parametrize("dataset", ["NGC6440E", "sim_sw.wb"])