+++ src/model/timing_model.jl
@@ -60,6 +60,7 @@ read_params(model::TimingModel, values) = read_params(model.param_handler, value
 get_free_param_names(model::TimingModel) = get_free_param_names(model.param_handler)
 get_free_param_units(model::TimingModel) = get_free_param_units(model.param_handler)
 get_free_param_labels(model::TimingModel) = get_free_param_labels(model.param_handler)
+get_free_param_prefixes(model::TimingModel) = get_free_param_prefixes(model.param_handler)
 
 read_param_values_to_vector(model::TimingModel) =
     read_param_values_to_vector(model.param_handler)